dbx-app-runner/pyproject.toml
[project]
name = "dbx-app-runner"
version = "0.0.1+ged195ca"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dynaconf",
    "giturlparse",
    "psutil",
    "pyaml",
    "dbx-core @ file://${PROJECT_ROOT}/../dbx-core",
    "dbx-tools @ file://${PROJECT_ROOT}/../dbx-tools",
    "requirements-parser",
    "sh",
]
[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources.dbx-core]
workspace = true

[tool.uv.sources.dbx-tools]
workspace = true
